Feature/initial logics

+++ src/main/java/com/tech/challenge/parking/meter/api/domain/dto/response/ParquimetroResponseDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class ParquimetroResponseDTO {
     private BigDecimal precoInicial;
-    private BigDecimal horasPrecoInicial;
+    private Integer horasPrecoInicial;
     private BigDecimal precoHoraExtra;
     private Integer totalVagas;
 }
